+++ ci-internal/DevDeployJobJenkinsfile.groovy
@@ -109,9 +109,7 @@ pipeline.stages = [
             script.sh("./gradlew writeToFileReleaseNotesDiff -PrevisionToCompare=${prevCommitHash}")
             String releaseNotesChanges = script.readFile(releaseNotesChangesFileUrl)
 
-            script.echo "qaas $releaseNotesChanges"
-
-//            JarvisUtil.sendMessageToGroup(script, releaseNotesChanges, idChatAndroidSlack, "slack", true)
+            JarvisUtil.sendMessageToGroup(script, " $releaseNotesChanges", idChatAndroidSlack, "slack", true)
         },
         pipeline.stage(CHECK_BRANCH_AND_VERSION) {
             String globalConfigurationJsonStr = script.readFile(projectConfigurationFile)