+++ easyadapter/lib-easyadapter/src/main/java/ru/surfstudio/android/easyadapter/diff/async/base/BaseAsyncDiffer.kt
@@ -15,10 +15,12 @@
  */
 package ru.surfstudio.android.easyadapter.diff.async.base
 
+import android.os.Handler
 import ru.surfstudio.android.easyadapter.diff.base.BaseDiffer
 import ru.surfstudio.android.easyadapter.diff.base.DiffCallbackCreator
 import ru.surfstudio.android.easyadapter.diff.base.DiffResultApplier
 import ru.surfstudio.android.easyadapter.diff.base.data.DiffCalculationBundle
+import ru.surfstudio.android.easyadapter.diff.base.data.DiffResultBundle
 import java.util.concurrent.Executors
 
 /**
@@ -29,12 +31,18 @@ internal abstract class BaseAsyncDiffer(
         diffCallbackCreator: DiffCallbackCreator
 ) : BaseDiffer(diffResultApplier, diffCallbackCreator), AsyncDiffer {
 
+    private val handler = Handler()
+
     private val diffExecutor = Executors.newFixedThreadPool(DIFF_EXECUTOR_POOL_SIZE)
 
     override fun startDiffCalculation(diffCalculationBundle: DiffCalculationBundle) {
         diffExecutor.execute {
             computeDiff(diffCalculationBundle)
         }
+    }
+
+    override fun dispatchDiffResult(diffResultBundle: DiffResultBundle) {
+        handler.post { applyDiffResult(diffResultBundle) }
     }
 
     private companion object {