+++ logger/lib-logger/src/main/java/ru/surfstudio/android/logger/remote_logging_strategies/impl/crashlytics/CrashlyticsRemoteLoggingStrategy.kt
@@ -1,3 +1,18 @@
+/*
+  Copyright (c) 2020-present, SurfStudio LLC.
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
 package ru.surfstudio.android.logger.remote_logging_strategies.impl.crashlytics
 
 import com.crashlytics.android.Crashlytics
@@ -5,8 +20,9 @@ import com.crashlytics.android.Crashlytics
 import ru.surfstudio.android.logger.remote_logging_strategies.RemoteLoggingStrategy
 
 /**
- * Стратегия логгирования в Crashlytics
+ * Logging strategy for Fabric Crashlytics
  */
+@Deprecated("Use FirebaseCrashlyticsRemoteLoggingStrategy and firebase crashlytics instead")
 class CrashlyticsRemoteLoggingStrategy : RemoteLoggingStrategy {
 
     override fun setUser(id: String?, username: String?, email: String?) {