+++ logger/RELEASE_NOTES.md
@@ -6,3 +6,5 @@
 ## 0.3.0
 ##### Logger
 * ANDDEP-222 - added the ability to select a strategy for logging
+* ANDDEP-392 - Add FirebaseCrashlyticsRemoteLoggingStrategy
+* ANDDEP-927 - Remove Firebase dependencies