+++ logger/RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 # Logger Release Notes
 ## 0.5.0-alpha
 ##### Logger
-* TODO
+* ANDDEP-950 Add different gradle script for modules without deploy
 ## 0.3.0
 ##### Logger
 * ANDDEP-222 - added the ability to select a strategy for logging