+++ logger/RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 ##### Logger
 * ANDDEP-392 - Add FirebaseCrashlyticsRemoteLoggingStrategy
 * ANDDEP-927 - Remove Firebase dependencies
-* ANDDEP-928 - Remove Crashlytics dependencies## 0.3.0
+* ANDDEP-928 - Remove Crashlytics dependencies
+## 0.3.0
 ##### Logger
 * ANDDEP-222 - added the ability to select a strategy for logging